+++ components/amorphic/lib/types/IAmorphicRouteController.d.ts
@@ -9,7 +9,7 @@ import { ControllerResponse } from './HelperTypes';
  * 
  * @misnomer - Functionality that may have confusing nomenclature
  * 
- * @deprecated - Deprecated functionality that is not used anymore
+ * @ deprecated - Deprecated functionality that is not used anymore
  */
 
 /**