+++ components/amorphic/lib/types/remote.d.ts
@@ -31,6 +31,12 @@ type RemoteDecoratorProps = {
 	serverValidation?: ServerValidationFunction;
 
 	/**
+	 * Lets you mark this function as public. By default it is false. When a remote call to this function is triggered,
+	 * the isPublic parameter in the preServerCall will reflect the public status of the function
+	 */
+	public?: boolean;
+
+	/**
 	 * A synchronous client side validation callback called in the case where code in the browser calls code in the server
 	 * and the browser code wishes to validate the input before making the call and throw an error
 	 *