+++ components/semotus/HISTORY.md
@@ -1,3 +1,8 @@
+## 4.1.1
+* Fixing a bug in semotus where checking to see if `public: true` would fail if the remote object instance did not exist within the session
+## 4.1.0
+* Added Semotus changes for `public: true` tag in remote decorator to help define public facing routes
+* Added ability for amorphic to pass in the request and response objects into semotus so it can pass them into the preserver and postServercall
 ## 4.0.0
 * Adding Sync State feature to limit object synchronization at stages in a session. See documentation in SyncStates.md
 ## 3.3.0